# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Education for Unprivileged Communities in India API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StatsSummary {
  outOfSchoolChildren: number;
  dropoutRateElementary: number;
  adolescentsNotFinishingSecondary: number;
  gdpSpendOnEducation: number;
  schoolsWithInternet: number;
  trainedTeachers: number;
  adultLiteracyClasses: number;
}

export type StateEducationStatRegion =
  (typeof StateEducationStatRegion)[keyof typeof StateEducationStatRegion];

export const StateEducationStatRegion = {
  north: "north",
  south: "south",
  east: "east",
  west: "west",
  central: "central",
  northeast: "northeast",
} as const;

export interface StateEducationStat {
  state: string;
  /** Name matching GeoJSON NAME_1 property */
  geoName: string;
  region: StateEducationStatRegion;
  /** Elementary dropout rate % (UDISE+ 2022-23) */
  dropoutRate: number;
  /** Primary Gross Enrollment Ratio % */
  primaryGer: number;
  /** Secondary Gross Enrollment Ratio % */
  secondaryGer: number;
  /** Overall literacy rate % (Census 2011 / NFHS-5) */
  literacyRate: number;
  /** Estimated out-of-school children */
  outOfSchoolChildren: number;
  /** Girls-specific elementary dropout rate % */
  girlsDropoutRate: number;
  /** One notable fact or program about the state */
  highlight: string;
}

export interface DropoutRate {
  group: string;
  rate: number;
  description: string;
}

export interface EnrollmentTrend {
  year: string;
  primaryEnrollment: number;
  secondaryEnrollment: number;
  higherEnrollment: number;
}

export type BarrierSeverity =
  (typeof BarrierSeverity)[keyof typeof BarrierSeverity];

export const BarrierSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Barrier {
  id: number;
  category: string;
  title: string;
  description: string;
  affectedGroups: string[];
  severity: BarrierSeverity;
}

export type InterventionLevel =
  (typeof InterventionLevel)[keyof typeof InterventionLevel];

export const InterventionLevel = {
  early_childhood: "early_childhood",
  primary: "primary",
  secondary: "secondary",
  vocational: "vocational",
  adult: "adult",
  digital: "digital",
} as const;

export type InterventionEvidenceLevel =
  (typeof InterventionEvidenceLevel)[keyof typeof InterventionEvidenceLevel];

export const InterventionEvidenceLevel = {
  pilot: "pilot",
  moderate: "moderate",
  strong: "strong",
  rigorous_rct: "rigorous_rct",
} as const;

export interface InstitutionLink {
  name: string;
  url: string;
}

export interface Intervention {
  id: number;
  name: string;
  level: InterventionLevel;
  description: string;
  impact: string;
  organization: string;
  organizationUrl: string;
  organizationLinks: InstitutionLink[];
  evidenceLevel: InterventionEvidenceLevel;
}

export interface CaseStudy {
  id: number;
  title: string;
  location: string;
  state: string;
  organization: string;
  organizationUrl: string;
  targetGroup: string;
  model: string;
  /** @nullable */
  costPerBeneficiary: string | null;
  outcomes: string;
  yearsActive: string;
  fundingSource: string;
  fundingLinks: InstitutionLink[];
}

export type FundingMechanismType =
  (typeof FundingMechanismType)[keyof typeof FundingMechanismType];

export const FundingMechanismType = {
  government: "government",
  csr: "csr",
  ngo: "ngo",
  impact_bond: "impact_bond",
  international_aid: "international_aid",
  mixed: "mixed",
} as const;

export interface FundingMechanism {
  id: number;
  name: string;
  type: FundingMechanismType;
  description: string;
  examples: string[];
  scale: string;
  websiteUrl: string;
  institutionLinks: InstitutionLink[];
}

export interface DonationRequest {
  name: string;
  email: string;
  amount: number;
  currency: string;
  /** @nullable */
  message?: string | null;
}

export interface DonationResponse {
  id: number;
  message: string;
  name: string;
  amount: number;
}

export type ContactRequestType =
  (typeof ContactRequestType)[keyof typeof ContactRequestType];

export const ContactRequestType = {
  volunteer: "volunteer",
  partner: "partner",
  researcher: "researcher",
  media: "media",
  other: "other",
} as const;

export interface ContactRequest {
  name: string;
  email: string;
  subject: string;
  message: string;
  type: ContactRequestType;
}

export interface ContactResponse {
  id: number;
  message: string;
}

export type InstitutionType =
  (typeof InstitutionType)[keyof typeof InstitutionType];

export const InstitutionType = {
  "central-university": "central-university",
  "state-university": "state-university",
  iit: "iit",
  nit: "nit",
  iim: "iim",
  aiims: "aiims",
  iiit: "iiit",
  nlaw: "nlaw",
  cfti: "cfti",
  "deemed-university": "deemed-university",
  "du-college": "du-college",
  "mu-college": "mu-college",
  "sppu-college": "sppu-college",
  other: "other",
} as const;

export type ExamCategory = (typeof ExamCategory)[keyof typeof ExamCategory];

export const ExamCategory = {
  engineering: "engineering",
  medical: "medical",
  management: "management",
  law: "law",
  design: "design",
  sciences: "sciences",
  "civil-services": "civil-services",
} as const;

export type ExamLevel = (typeof ExamLevel)[keyof typeof ExamLevel];

export const ExamLevel = {
  ug: "ug",
  pg: "pg",
  both: "both",
} as const;

export type ExamMode = (typeof ExamMode)[keyof typeof ExamMode];

export const ExamMode = {
  online: "online",
  offline: "offline",
  both: "both",
} as const;

export type ExamFrequency = (typeof ExamFrequency)[keyof typeof ExamFrequency];

export const ExamFrequency = {
  annual: "annual",
  biannual: "biannual",
} as const;

export interface Exam {
  id: string;
  name: string;
  shortName: string;
  category: ExamCategory;
  level: ExamLevel;
  conductingBody: string;
  conductingBodyUrl: string;
  officialWebsite: string;
  description: string;
  eligibility: string;
  applicationStart: string;
  applicationEnd: string;
  examDate: string;
  resultDate: string;
  /** @nullable */
  totalSeats: number | null;
  /** @nullable */
  candidatesAppearing: string | null;
  mode: ExamMode;
  frequency: ExamFrequency;
  eligibleInstitutionTypes: string[];
  acceptedBy: string[];
}

export interface Institution {
  id: string;
  name: string;
  shortName?: string;
  type: InstitutionType;
  state: string;
  city: string;
  established: number;
  website: string;
  naacGrade?: string;
  affiliation: string;
  description: string;
  courses: string[];
  ugcRecognized: boolean;
  nirfRank?: number;
}