# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Education for Unprivileged Communities in India API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /donate:
    post:
      operationId: submitDonation
      tags: [donate]
      summary: Submit a donation or support pledge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationRequest"
      responses:
        "201":
          description: Donation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationResponse"
        "400":
          description: Invalid input
  /contact:
    post:
      operationId: submitContact
      tags: [donate]
      summary: Submit a contact or volunteer inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
      responses:
        "201":
          description: Contact recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Invalid input
...
components:
  schemas:
...
    DonationRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        amount:
          type: number
        currency:
          type: string
        message:
          type: ["string", "null"]
      required:
        - name
        - email
        - amount
        - currency
    DonationResponse:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        name:
          type: string
        amount:
          type: number
      required:
        - id
        - message
        - name
        - amount
    ContactRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [volunteer, partner, researcher, media, other]
      required:
        - name
        - email
        - subject
        - message
        - type
    ContactResponse:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
      required:
        - id
        - message
...